backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo==4.6.2
mongomock==4.1.2
python-dotenv==1.0.0

